Procedure to configure a KDENLIVE project dataset
This new dataset procedure configures a dataset for use as a KDENLIVE
project. Most importantly it configures the dataset clone to not commit
the absolute root path configured in a project file into Git. Instead, a
placeholder is stored that is replaced with the absolute path of the
dataset root location on checkout. This makes KDENLIVE project datasets
portable.

Moreover, it:

- expects the project file to be named 'project.kdenlive', and to be
  placed in the root of the dataset
- creates a standard directory structure
- configures Git to ignore a number of temporary working directories
  that KDENLIVE might create

The dataset procedure is idempotent, and is save to run on
datalad-create (with -c), and must be executed once in any clone
of such a dataset to apply all necessary configuration.