More accurate handling of vanished remote states and upload failures
The local mirror repo is now more strictly sync'ed to the remote state,
and is wiped out when no remote state is detected any longer. Previously
the mirror was kept in this case. This lead to confusing messages like

 = [up to date]      master -> master

when uploading anew, and not actually being truly up-to-date. With this
change it would now message:

 * [new branch]      master -> master

This change also impacts the need to maintain the refs in the mirror
repo. Rollbacks in case of failed uploads are no longer necessary.
A failed upload will result in a check for remote and mirror being
in sync, and a replacement of the entire mirror if that is not the case.

As a related change the indicator approach to failed uploads was
simplified. Instead of annotating the last confirmed remote state
for individual refs the were pushed, now the actual pushed state
is annotated with a `dlra-upload-failed` refs. These markers
are now all removed wit the next successful upload, not just the
particular refs being pushed. This avoid accumulation of cruft. It is
possible, because always a complete remote state is uploaded, and a
successful upload means all refs that should be present at a remote,
are indeed present.