Simplified metadata homogenization approach
This approach first takes a document, and converts it to triples.
Triples are the base unit of a graph and have a simple and predictable
structure.

Term mapping for homogenization is applied to any IRI
and data type declaration. Terms that are not included in a given
term mapping are left as-is.

Here is how this can be used in an application (here generate a document
on a dataset and its contained files, with a simple yet precisely defined
structure).

We start with https://raw.githubusercontent.com/psychoinformatics-de/datalad-tabby/8f411d87b1a2d216e10b8c9aa5d18f91d2df444e/docs/source/conventions/tby-ds1_demo/compacted.jsonld

```py
>>> import datalad_mihextras.jsonld as jd
>>> import json
>>> from pathlib import Path
>>> jdoc = json.load(Path('/tmp/compacted.jsonld').open())
>>> homogenized = jd.homogenize(jdoc, jd.equivalence_map, jd.compaction_context)
```

We can now assume a particular terminology to be used in the document
(obviously subject to the term mapping being adequate for the input and
the target). This allows for declaring a relatively simple JSON-LD frame

```json
{
    "@context": {
        "schema": "https://schema.org/",
        "md5sum": "https://purl.obolibrary.org/obo/NCIT_C171276",
        "url": "schema:contentUrl",
        "hasPart": "https://purl.org/dc/terms/hasPart",
        "name": "schema:name",
        "posixpath": "http://purl.allotrope.org/ontologies/result#AFR_0001928",
        "bytesize": "https://www.semanticdesktop.org/ontologies/2007/03/22/nfo/#fileSize",
    },
    "@type": "schema:Dataset",
    "hasPart": {
        "@type": "schema:DigitalDocument",
    }
}
```

This says: We want to have a `Dataset` object (one or more) at the top
of the framed document. This dataset must have `DigitalDocument`-type
objects linked to it via a `hasPart` relationship. The rest of the
document declared the desired JSON keys in a context, based on their
(homogenized) definitions.

```py
jd.jsonld.frame(homogenized, jd.dsfiles_frame)
```

Now gives:

```json
{
  "@context": {
    "schema": "https://schema.org/",
    "md5sum": "https://purl.obolibrary.org/obo/NCIT_C171276",
    "url": "schema:contentUrl",
    "hasPart": "https://purl.org/dc/terms/hasPart",
    "name": "schema:name",
    "posixpath": "http://purl.allotrope.org/ontologies/result#AFR_0001928",
    "bytesize": "https://www.semanticdesktop.org/ontologies/2007/03/22/nfo/#fileSize"
  },
  "@type": "schema:Dataset",
  "hasPart": [
    {
      "@type": "schema:DigitalDocument",
      "md5sum": "529ff606a38b37a2e5478c1abfeca231",
      "url": "https://raw.githubusercontent.com/psychoinformatics-de/datalad-tabby/2738d8a12fb138d3fe107c6bee443c13c9f4f6ea/LICENSE",
      "name": {
        "@type": "posixpath",
        "@value": "LICENSE"
      },
      "bytesize": 1300
    },
    {
      "@type": "schema:DigitalDocument",
      "md5sum": "ef2979a70a8d95a24cd1402bd68e1c4a",
      "url": "https://raw.githubusercontent.com/psychoinformatics-de/datalad-tabby/2738d8a12fb138d3fe107c6bee443c13c9f4f6ea/docs/README.md",
      "name": {
        "@type": "posixpath",
        "@value": "docs/README.md"
      },
      "bytesize": 1755
    }
  ],
  "schema:author": {
    "@type": "schema:Person",
    "schema:email": "jd@example.com",
    "name": "Jane Doe"
  },
  "schema:dateModified": "2023-07-27",
  "schema:description": "This is a fictitious dataset.",
  "schema:license": {
    "@id": "https://spdx.org/licenses/CC-PDDC"
  },
  "schema:mainEntityOfPage": "http://docs.datalad.org/projects/tabby/en/latest",
  "name": "demo",
  "schema:title": "My demo dataset"
}
```

which is relatively simple to process, ie. for compiling a
`datalad-addurls` call from the `hasPart` list.

Also note that the path type is pretty explicit, and that the file size
is given in JSON-native integer values.

Yes, this example does not actually do effective term mapping. However,
any match in `jd.equivalence_map` would get mapped, and any
homogenization of terms boils down to assembling such a mapping. The
rest is done by this code.

Ping #40